my updated express project

+++ controllers/clothingItem.js
@@ -1,36 +1,43 @@
 const clothingItem = require("../models/clothingItem");
 const ClothingItem = require("../models/clothingItem");
-const { INTERNAL_SERVER_ERROR_CODE } = require("../utils/errors");
+const {
+  INTERNAL_SERVER_ERROR_CODE,
+  NOT_FOUND_ERROR_CODE,
+} = require("../utils/errors");
+const { BAD_REQUEST_ERROR_CODE } = require("../utils/errors");
 
-//main handler of all of my api calls. Line 5
-//console.log(req.body) is when we work with a post. The "body" contains the majority of the data
-//next, we extract the data out of the body on line 10. Start with const then the date in {}
-//next, ClothingItem.create({name, wealther, imageURL}) is apart of the express
-//ClothingItem.create({name, wealther, imageURL}) is a promise so use .then and return ((item))
+// main handler of all of my api calls. Line 5
+// console.log(req.body) is when we work with a post. The "body" contains the majority of the data
+// next, we extract the data out of the body on line 10. Start with const then the date in {}
+// next, ClothingItem.create({name, wealther, imageURL}) is apart of the express
+// ClothingItem.create({name, wealther, imageURL}) is a promise so use .then and return ((item))
 const createItem = (req, res) => {
-  console.log(req);
-  console.log(req.body);
+  const { name, weather, imageUrl } = req.body;
 
-  const { name, wealther, imageURL } = req.body;
-
-  ClothingItem.create({ name, wealther, imageURL })
+  ClothingItem.create({ name, weather, imageUrl })
     .then((item) => {
-      console.log("item");
       res.send({ data: item });
     })
     .catch((e) => {
+      if (e.name === "ValidationError") {
+        return res
+          .status(BAD_REQUEST_ERROR_CODE)
+          .send({ message: "Invalid user ID" });
+      }
+      console.error(e);
       res
         .status(INTERNAL_SERVER_ERROR_CODE)
         .send({ message: "Error from createItem", e });
     });
 };
 
-//This function connects to the "Read" route in clothingItem.js
+// This function connects to the "Read" route in clothingItem.js
 const getItems = (req, res) => {
   clothingItem
     .find({})
     .then((items) => res.status(200).send(items))
     .catch((e) => {
+      console.error(e);
       res
         .status(INTERNAL_SERVER_ERROR_CODE)
         .send({ message: "Error from getItems", e });
@@ -48,6 +55,7 @@ const updateItem = (req, res) => {
       res.status(200).send({ data: item });
     })
     .catch((e) => {
+      console.error(e);
       res
         .status(INTERNAL_SERVER_ERROR_CODE)
         .send({ message: "Error from updateItem", e });
@@ -63,12 +71,79 @@ const deleteItem = (req, res) => {
     .findByIdAndDelete(itemId)
     .orFail()
     .then(() => {
-      res.status(204).send({ message: "Item deleted successfully" });
+      res.status(200).send({ message: "Item deleted successfully" });
     })
     .catch((e) => {
+      if (e.name === "DocumentNotFoundError") {
+        return res
+          .status(NOT_FOUND_ERROR_CODE)
+          .send({ message: "Item Not Found" });
+      }
+      if (e.name === "CastError") {
+        return res
+          .status(BAD_REQUEST_ERROR_CODE)
+          .send({ message: "Invalid item ID" });
+      }
+      console.error(e);
       res
         .status(INTERNAL_SERVER_ERROR_CODE)
         .send({ message: "Error from deleteItem", e });
+    });
+};
+
+const likeItem = (req, res) => {
+  const { itemId } = req.params;
+  const userId = req.user._id;
+
+  clothingItem
+    .findByIdAndUpdate(itemId, { $addToSet: { likes: userId } }, { new: true })
+    .orFail()
+    .then((item) => {
+      res.status(200).send({ data: item });
+    })
+    .catch((e) => {
+      if (e.name === "DocumentNotFoundError") {
+        return res
+          .status(NOT_FOUND_ERROR_CODE)
+          .send({ message: "Item Not Found" });
+      }
+      if (e.name === "CastError") {
+        return res
+          .status(BAD_REQUEST_ERROR_CODE)
+          .send({ message: "Invalid item ID" });
+      }
+      console.error(e);
+      res
+        .status(INTERNAL_SERVER_ERROR_CODE)
+        .send({ message: "Error from likeItem", e });
+    });
+};
+
+const unlikeItem = (req, res) => {
+  const { itemId } = req.params;
+  const userId = req.user._id;
+
+  clothingItem
+    .findByIdAndUpdate(itemId, { $pull: { likes: userId } }, { new: true })
+    .orFail()
+    .then((item) => {
+      res.status(200).send({ data: item });
+    })
+    .catch((e) => {
+      if (e.name === "DocumentNotFoundError") {
+        return res
+          .status(NOT_FOUND_ERROR_CODE)
+          .send({ message: "Item Not Found" });
+      }
+      if (e.name === "CastError") {
+        return res
+          .status(BAD_REQUEST_ERROR_CODE)
+          .send({ message: "Invalid item ID" });
+      }
+      console.error(e);
+      res
+        .status(INTERNAL_SERVER_ERROR_CODE)
+        .send({ message: "Error from unlikeItem", e });
     });
 };
 
@@ -77,4 +152,6 @@ module.exports = {
   getItems,
   updateItem,
   deleteItem,
+  likeItem,
+  unlikeItem,
 };
